ファイルの添付とDatePickerを共通化した
