+++ BookManager/AddBookViewController.swift
@@ -21,22 +21,12 @@ class AddBookViewController: UIViewController, UIImagePickerControllerDelegate, 
     }
 
     @IBAction func popupDatepickerAction(_ sender: UITextField) {
-        let datePicker = UIDatePicker()
-        datePicker.datePickerMode = UIDatePickerMode.date
-        sender.inputView = datePicker
-        datePicker.addTarget(self, action: #selector(pickerChanged),
-                             for: UIControlEvents.valueChanged)
+        DatePickerHelper.load(self, action: #selector(pickerChanged), sender: sender)
+
     }
 
     @IBAction func attachImageAction(_ sender: Any) {
-        let photo = UIImagePickerControllerSourceType.photoLibrary
-
-        if UIImagePickerController.isSourceTypeAvailable(photo) {
-            let picker = UIImagePickerController()
-            picker.sourceType = photo
-            picker.delegate = self
-            self.present(picker, animated: true)
-        }
+        FileAttachHelper.load(self, delegate: self)
     }
 
     @IBAction func saveAction(_ sender: Any) {
